Added simple Dockerfile for building the kong package This uses the same ubuntu-hardened-fips image that we target to run it on
Reverted OpenSSL version that bazel builds & links openresty against to 1.1.1. This is the fips-validated version available in our ubuntu image.

Changed final bazel build to NOT copy its compiled version of openssl to the final output. This ensures it won't be used by accident and instead the OS-provided version will be leveraged.